refactor : refactor file structure
