+++ login-svc/app/api/__init__.py
@@ -1,0 +1,3 @@
+# app/api/__init__.py
+from .auth import router as auth_router
+from .users import router as users_router
