+++ login-svc/app/models/user.py
@@ -1,7 +1,9 @@
 from sqlalchemy import Column, Integer, String
 from app.database import Base
 
+# user db models 
 class User(Base):
+    # 表格的名稱是 users
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True, index=True)
