+++ login-svc/app/repositories/__init__.py
@@ -1,0 +1,2 @@
+from .repository_mock import MockUserRepository
+from .repository import UserRepository
