+++ login-svc/app/schemas/__init__.py
@@ -1,0 +1,3 @@
+from .schemas import UserBase, UserCreate, UserResponse, LoginRequest, Token
+
+__all__ = ["UserBase", "UserCreate", "UserResponse", "LoginRequest", "Token"]
